pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "driftbench"
version = "0.1.0"
description = "DriftBench: measuring LLM safety-rule drift over long conversations"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Benji Berczi" }
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["models", "builders", "scoring", "metrics", "analysis"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
